audio stuff
Collision audio and police radio stuff moved to a separate cpp's like in the original code
cPoliceRadioQueue added, code refactored correspondingly
cAudioManager::ServiceCollisions reversed
PlayOneShotScriptObject reversed
cAudioScriptObject moved to AudioScriptObject.cpp
Fixed use of numbers on non-integer types (bools, pointers)
enum is now used for ped audio calculation
Added fixes for vehicle description in the police radio (missing vehicles and wrong colors)
