+++ src/audio/audio_enums.h
@@ -226,3 +226,30 @@ enum eStreamedSounds
 	TOTAL_STREAMED_SOUNDS = 196,
 	NO_STREAMED_SOUND = 197,
 };
+
+enum AudioEntityHandle {
+	AEHANDLE_NONE = -5,
+	AEHANDLE_ERROR_NOAUDIOSYS = -4,
+	AEHANDLE_ERROR_NOFREESLOT = -3,
+	AEHANDLE_ERROR_NOENTITY = -2,
+	AEHANDLE_ERROR_BADAUDIOTYPE = -1,
+};
+
+enum eAudioType : int32
+{
+	AUDIOTYPE_PHYSICAL = 0,
+	AUDIOTYPE_EXPLOSION = 1,
+	AUDIOTYPE_FIRE = 2,
+	AUDIOTYPE_WEATHER = 3,
+	AUDIOTYPE_CRANE = 4,
+	AUDIOTYPE_SCRIPTOBJECT = 5,
+	AUDIOTYPE_BRIDGE = 6,
+	AUDIOTYPE_COLLISION = 7,
+	AUDIOTYPE_FRONTEND = 8,
+	AUDIOTYPE_PROJECTILE = 9,
+	AUDIOTYPE_GARAGE = 10,
+	AUDIOTYPE_FIREHYDRANT = 11,
+	AUDIOTYPE_WATERCANNON = 12,
+	AUDIOTYPE_POLICERADIO = 13,
+	TOTAL_AUDIO_TYPES = 14,
+};
