+++ src/core/Wanted.h
@@ -22,6 +22,7 @@ enum eCrimeType
 	CRIME_COP_BURNED,
 	CRIME_VEHICLE_BURNED,
 	CRIME_DESTROYED_CESSNA,
+	NUM_CRIME_TYPES
 };
 
 class CCrimeBeingQd
